Get tests to work without the massive project structure change.
