+++ test/tables/test_flags.py
@@ -1,5 +1,5 @@
 import pytest
-from pipeline.src.tables.table import Table
+from src.tables.table import Table
 
 def test_hello_world():
     print("Hello, world!")
